+++ experimental/packages/opentelemetry-instrumentation-http/test/functionals/utils.test.ts
@@ -19,6 +19,7 @@ import {
   SpanKind,
   context,
   Span,
+  diag,
 } from '@opentelemetry/api';
 import {
   SEMATTRS_HTTP_REQUEST_CONTENT_LENGTH,
@@ -101,7 +102,7 @@ describe('Utility', () => {
         urlParsedWithUndefinedHostAndNullPort,
         whatWgUrl,
       ]) {
-        const result = utils.getRequestInfo(param);
+        const result = utils.getRequestInfo(diag, param);
         assert.strictEqual(result.optionsParsed.hostname, 'google.fr');
         assert.strictEqual(result.optionsParsed.protocol, 'http:');
         assert.strictEqual(result.optionsParsed.path, '/aPath?qu=ry');
@@ -432,10 +433,14 @@ describe('Utility', () => {
         'user-agent': 'chrome',
         'x-forwarded-for': '<client>, <proxy1>, <proxy2>',
       };
-      const attributes = utils.getIncomingRequestAttributes(request, {
-        component: 'http',
-        semconvStability: SemconvStability.OLD,
-      });
+      const attributes = utils.getIncomingRequestAttributes(
+        request,
+        {
+          component: 'http',
+          semconvStability: SemconvStability.OLD,
+        },
+        diag
+      );
       assert.strictEqual(attributes[SEMATTRS_HTTP_ROUTE], undefined);
     });
 
@@ -448,10 +453,14 @@ describe('Utility', () => {
       request.headers = {
         'user-agent': 'chrome',
       };
-      const attributes = utils.getIncomingRequestAttributes(request, {
-        component: 'http',
-        semconvStability: SemconvStability.OLD,
-      });
+      const attributes = utils.getIncomingRequestAttributes(
+        request,
+        {
+          component: 'http',
+          semconvStability: SemconvStability.OLD,
+        },
+        diag
+      );
       assert.strictEqual(attributes[SEMATTRS_HTTP_TARGET], '/user/?q=val');
     });
   });